Rename type file for size mapping
